core: move solving hashcash into util
